add support for auto-discovering node group options from the cloud provider
This adds support for auto-discovering min_nodes and max_nodes from the cloud provider
for each node group. The auto-discovering/setting of the values is only done on the
initial start of Escalator. It can be enabled/disabled by setting both min_nodes and
max_nodes to 0 or removing them from the node groups config.

fixes #121